create custom plugin dependency library
