+++ metadata-io/build.gradle
@@ -10,6 +10,7 @@ configurations {
 dependencies {
   implementation project(':entity-registry')
   implementation project(':metadata-service:auth-config')
+  api project(':metadata-io:metadata-io-api')
   api project(':metadata-utils')
   api project(':metadata-events:mxe-avro')
   api project(':metadata-events:mxe-registration')
