+++ metadata-io/metadata-io-api/README.md
@@ -1,0 +1,4 @@
+# :metadata-io:metadata-io-api
+
+This module exists in order to isolate dependencies when used in external projects. For example,
+a custom plugin implementing a custom validator, mutator, or side-effect.
