+++ settings.gradle
@@ -43,6 +43,7 @@ include 'metadata-testing:metadata-models-test-utils'
 include 'metadata-testing:metadata-test-utils'
 include 'entity-registry'
 include 'metadata-io'
+include 'metadata-io:metadata-io-api'
 include 'datahub-upgrade'
 include 'metadata-utils'
 include 'li-utils'
@@ -53,6 +54,7 @@ include 'metadata-models-custom'
 include 'entity-registry:custom-test-model'
 include 'metadata-integration:java:spark-lineage'
 include 'metadata-integration:java:datahub-client'
+include 'metadata-integration:java:custom-plugin-lib'
 include 'metadata-integration:java:datahub-event'
 include 'metadata-integration:java:datahub-protobuf'
 include 'metadata-integration:java:openlineage-converter'
